store: Explicitly cast struct timeval values to long int for printf
Otherwise we are depending on them being defined as long int anyway in
order to get them to be passed correctly so we might as well just
explicitly cast them.

+++ src/ntb-store.c
@@ -860,8 +860,8 @@ generate_maildir_name(struct ntb_store *store,
 
         ntb_buffer_append_printf(buffer,
                                  "%li.M%liQ%u.",
-                                 tv.tv_sec,
-                                 tv.tv_usec,
+                                 (long int) tv.tv_sec,
+                                 (long int) tv.tv_usec,
                                  store->num_stored_messages++);
 
         while (true) {
